+++ cli/commands/__init__.py
@@ -1,0 +1,7 @@
+import argparse
+
+from cli.commands.compile import parse_compile_args, run_compile
+from cli.commands.convert import parse_convert_args, run_convert
+from cli.commands.simulate import parse_simulate_args, run_simulate
+
+__all__ = ["parse_convert_args", "run_convert", "parse_compile_args", "run_compile", "parse_simulate_args", "run_simulate"]
